1808 JUnit review

+++ src/main/java/com/revature/screenforce/controllers/QuestionController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.revature.screenforce.beans.Question;
+import com.revature.screenforce.services.BucketService;
 import com.revature.screenforce.services.QuestionService;
 
 import javax.validation.Valid;
@@ -37,6 +38,9 @@ public class QuestionController {
 
 	@Autowired
 	private QuestionService qs;
+	
+	@Autowired
+	private BucketService bs;
 
 	/**
 	 * Get all questions
@@ -46,7 +50,7 @@ public class QuestionController {
 			response = Question.class,
 		    responseContainer = "List")
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "All questions returned") } )
-	@GetMapping("/all")
+	@GetMapping()
 	public ResponseEntity<List<Question>> getQuestions() {
 		return new ResponseEntity<>(qs.getAllQuestions(), HttpStatus.OK);
 	}
@@ -76,23 +80,43 @@ public class QuestionController {
 	 * Find all questions associated with a specific bucket
 	 * @param bucketId Id of bucket to filter by
 	 * @return List of questions associated with bucket of given id
+	 * 
+	 * For the getByBucket method, it should be done using a params. rather than an endpoint. 
 	 */
 	@ApiOperation(value = "Find list of Questions by bucketId",
 		notes = "Each question belongs to a particular bucket that classifies subject matter",
 	    response = Question.class,
 	    responseContainer = "List")
-	@ApiResponses(value = { @ApiResponse(code = 200, message = "Response is empty if bucketId is not found") } )
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Success, bucketId found with Question Associated"),
+			@ApiResponse(code = 404, message = "bucketId is not found")} )
 	@GetMapping("/getByBucket/{bucketId}")
 	public ResponseEntity<List<Question>> getBucketQuestions(@PathVariable(value="bucketId") int bucketId) {
-		return new ResponseEntity<>(qs.getQuestionsByBucket(bucketId), HttpStatus.OK);
+		if (bs.existsById(bucketId)) {
+			return new ResponseEntity<>(qs.getQuestionsByBucket(bucketId), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
 	}
-
+	/**
+	 * 
+	 * @param bucketId
+	 * @return Void
+	 * 
+	 */
 	@ApiOperation(value = "Delete all questions associated with a bucketId")
-	@ApiResponse(code = 204, message = "Deleted associated questions")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "bucketId found, deleted all question"),
+			@ApiResponse(code = 404, message = "bucketId is not found")} )
 	@DeleteMapping("/deleteByBucket/{bucketId}")
 	public ResponseEntity<Void> deleteByBucket(@PathVariable(value = "bucketId") int bucketId){
-		qs.deleteByBucketId(bucketId);
-		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		if(bs.existsById(bucketId)) {
+			qs.deleteByBucketId(bucketId);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	/**
@@ -102,22 +126,28 @@ public class QuestionController {
 	 */
 	@ApiOperation(value = "Adds a new Question", response = Question.class)
 	@ApiResponses(value = { @ApiResponse(code = 201, message = "New question created") } )
-	@PostMapping(value = "/new" ,consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Question> create(@Valid @RequestBody Question question) {
 		return new ResponseEntity<>(this.qs.create(question), HttpStatus.CREATED);
 	}
 	
 	/**
 	 * Updates a number of fields for a specific question
-	 * @param question updated question object
+	 * @param updated question object
 	 * @return updated question and http status code 200
 	 */
 	@ApiOperation(value = "Updates question", response = Question.class)
-	@ApiResponses(value = { @ApiResponse(code = 200, message = "Question updated") } )
-	@PutMapping(value = "/update" , consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Question> updateQuestion(@Valid @RequestBody Question question) {
-		qs.updateQuestion(question);
-		return new ResponseEntity<>(question, HttpStatus.OK);
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Question updated"),
+			@ApiResponse(code = 400, message = "Bad Request, Question is not updated") } )
+	@PutMapping(value = "/{id}" , consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<Question> updateQuestion(@PathVariable(value = "id") int id, @RequestBody Question question) {
+		if (qs.existsById(id)) {
+			qs.updateQuestion(question);
+			return new ResponseEntity<>(question, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	/**
@@ -126,23 +156,16 @@ public class QuestionController {
 	 * @return http status code 200
 	 */
 	@ApiOperation(value = "Deletes a question")
-	@ApiResponses(value = { @ApiResponse(code = 200, message = "Question deleted") } )
-	@DeleteMapping("/delete/{id}")
-	public ResponseEntity<Void> deleteByQuestionId(@PathVariable(value="id") Integer questionId) {
-		qs.deleteByQuestionId(questionId);
-		return new ResponseEntity<>(HttpStatus.OK);
-	}
-	
-	/**
-	 * Toggles question isActive status
-	 * @param questionId - id of question to remove
-	 * @return http status code 204
-	 */
-	@ApiOperation(value = "Sets Question to active state", response = void.class)
-	@ApiResponses(value = { @ApiResponse(code = 204, message = "Question toggled") } )
-	@PutMapping("/toggle/{id}")
-	public ResponseEntity<Void> activateQuestion(@PathVariable(value="id") Integer questionId) {
-		qs.toggleQuestionStatus(questionId);
-		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Question deleted"),
+			@ApiResponse(code = 400, message = "Bad Request, QuestionID not found") } )
+	@DeleteMapping("/{id}")
+	public ResponseEntity<Void> deleteByQuestionId(@PathVariable(value="id") int id) {
+		if (qs.existsById(id)) {
+			qs.deleteByQuestionId(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 	}
 }
